split shop pages
