+++ client/src/components/shop/BrandsSearch.jsx
@@ -1,7 +1,38 @@
 import React from 'react'
-import {Radio} from 'antd'
+import { Radio } from 'antd'
+import { useDispatch } from 'react-redux'
 
-const BrandsSearch = ({brand, brands, handleBrand}) => {
+
+const BrandsSearch = ({
+  brand,
+  brands,
+  setPrice,
+  setStar,
+  setBrand,
+  setCategoryIds,
+  setSub,
+  setColor,
+  setShipping,
+  fetchProducts }) => {
+
+  let dispatch = useDispatch()
+  // 7 brandに基づいて商品を表示
+
+  const handleBrand = (e) => {
+    setSub('')
+    dispatch({
+      type: "SEARCH_QUERY",
+      payload: {text: ""}
+    })
+    setPrice([0,0])
+    setCategoryIds([])
+    setStar('')
+    setColor('')
+    setShipping('')
+    setBrand(e.target.value)
+    fetchProducts({brand: e.target.value})
+  }
+
   return (
     <div>
       {brands.map((b) =>
