+++ client/src/components/shop/CategoriesSearch.jsx
@@ -1,7 +1,50 @@
 import React from 'react'
+import { useDispatch } from 'react-redux'
 import { Checkbox } from 'antd';
 
-const CategoriesSearch = ({categories, handleCheck, categoryIds}) => {
+const CategoriesSearch = ({
+  categories,
+  categoryIds,
+  setPrice,
+  setStar,
+  setBrand,
+  setCategoryIds,
+  setSub,
+  setColor,
+  setShipping,
+  fetchProducts }) => {
+  let dispatch = useDispatch()
+  
+  // category に基づいて製品をload
+  const handleCheck = e => {
+    dispatch({
+      type: "SEARCH_QUERY",
+      payload: {text: ""}
+    })
+    setPrice([0,0])
+    setStar("")
+    setSub('')
+    setBrand("")
+    setColor('')
+    setShipping('')
+    //console.log(e.target.value);
+    let inTheState = [...categoryIds]
+    let justChecked = e.target.value
+    let foundInTheState = inTheState.indexOf(justChecked) //index or -1
+
+    //indexOf? 見つからない場合は-1を返し、そうでない場合はインデックスを返す
+    if (foundInTheState === -1) {
+      inTheState.push(justChecked)
+    } else {
+      // 見つかった場合は、インデックスから1つのアイテムを引き出す
+      inTheState.splice(foundInTheState, 1)
+    }
+    setCategoryIds(inTheState)
+    //console.log(inTheState);
+    fetchProducts({category: inTheState})
+  }
+
+
   return (
     <div>
       {categories.map((c) => (
