+++ client/src/components/shop/PriceSearch.jsx
@@ -1,8 +1,46 @@
-import React from 'react'
+import React, {useEffect} from 'react'
+import { useDispatch } from 'react-redux'
 import { Slider, } from 'antd';
 
 
-const PriceSearch = ({price, handleSlider}) => {
+const PriceSearch = ({
+  price,
+  ok,
+  setPrice,
+  setStar,
+  setBrand,
+  setCategoryIds,
+  setSub,
+  setColor,
+  setShipping,
+  setOk,
+  fetchProducts
+}) => {
+
+  let dispatch = useDispatch()
+
+  // priceに基づいて製品をload
+  useEffect(() => {
+    //console.log('price response');
+    fetchProducts({price})
+  }, [ok])
+
+  const handleSlider = (value) => {
+    dispatch({
+      type: "SEARCH_QUERY",
+      payload: {text: ""}
+    })
+    setStar("")
+    setBrand("")
+    setCategoryIds([])
+    setSub('')
+    setColor('')
+    setShipping('')
+    setPrice(value)
+    setTimeout(() => {
+      setOk(!ok)
+    },300)
+  }
   return (
     <Slider
       className="ml-4 mr-4"
