+++ client/src/components/shop/ShippingSearch.jsx
@@ -1,7 +1,34 @@
 import React from 'react'
-import {Checkbox} from 'antd'
+import { Checkbox } from 'antd'
+import { useDispatch } from 'react-redux'
 
-const ShippingSearch = ({handleShippingChange, shipping,}) => {
+const ShippingSearch = ({
+  setPrice,
+  setStar,
+  setBrand,
+  setCategoryIds,
+  setSub,
+  setColor,
+  setShipping,
+  fetchProducts,
+  shipping, }) => {
+  let dispatch = useDispatch()
+
+  //9 shippingに基づいて商品を表示
+  const handleShippingChange = (e) => {
+    setSub('')
+    dispatch({
+      type: "SEARCH_QUERY",
+      payload: {text: ""}
+    })
+    setPrice([0,0])
+    setCategoryIds([])
+    setStar('')
+    setBrand('')
+    setColor('')
+    setShipping(e.target.value)
+    fetchProducts({shipping: e.target.value})
+  }
   return (
     <>
       <Checkbox
