+++ client/src/components/shop/SubsSearch.jsx
@@ -1,6 +1,36 @@
 import React from 'react'
+import { useDispatch } from 'react-redux'
 
-const SubsSearch = ({subs, handleSubs}) => {
+const SubsSearch = ({
+  subs,
+  setPrice,
+  setStar,
+  setBrand,
+  setCategoryIds,
+  setSub,
+  setColor,
+  setShipping,
+  fetchProducts }) => {
+  
+  let dispatch = useDispatch()
+  
+  //6 sub category に基づいて商品を表示
+  const handleSubs = (sub) => {
+    //console.log(sub);
+    setSub(sub)
+    dispatch({
+      type: "SEARCH_QUERY",
+      payload: {text: ""}
+    })
+    setPrice([0,0])
+    setCategoryIds([])
+    setStar('')
+    setBrand("")
+    setColor('')
+    setShipping('')
+    fetchProducts({sub: sub})
+  }
+
   return (
     <div>
       {subs.map((s) =>
